Colin & Paddy & Fearghal - Code final
This is the final version of the app, tech report in the next push
