+++ 3. AudioAcousticAssistant-App/Php + Sql/Login.php
@@ -1,4 +1,6 @@
 <?php
+
+//@reference YouTube - TonikamiTV/Login Register 6 part series - https://www.youtube.com/watch?v=QxffHgiJ64M
 	//this imports the password library that does the encryption.
     require("password.php");
 	
@@ -24,7 +26,7 @@
     while(mysqli_stmt_fetch($statement)){
         if (password_verify($password, $colPassword)) {
             $response["success"] = true;  
-            //$response["email"] = $colEmail;
+            $response["email"] = $colEmail;
            
         }
     }
