+++ 3. AudioAcousticAssistant-App/app/src/main/AndroidManifest.xml
@@ -17,15 +17,16 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".RoomActivity" />
-        <activity android:name=".ChooseActivity" />
-        <activity android:name=".RecordActivity" />
-        <activity android:name=".AdviceActivity" />
+        <activity android:name=".RoomActivity" android:screenOrientation="portrait"/>
+        <activity android:name=".ChooseActivity" android:screenOrientation="portrait" />
+        <activity android:name=".RecordActivity" android:screenOrientation="portrait"/>
+        <activity android:name=".AdviceActivity" android:screenOrientation="portrait"/>
         <!--<activity android:name=".Splash" />-->
-        <activity android:name=".RegisterActivity" />
-        <activity android:name=".LoginActivity" />
-        <activity android:name=".SplashActivity"/>
-        <activity android:name=".MainActivity">
+        <activity android:name=".RegisterActivity" android:screenOrientation="portrait"/>
+        <activity android:name=".LoginActivity" android:screenOrientation="portrait"/>
+
+        <activity android:name=".MainActivity" android:screenOrientation="portrait"/>
+            <activity android:name=".SplashScreen" android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
