+++ 3. AudioAcousticAssistant-App/app/src/main/java/com/example/x15011071/audioacousticassistant_app/RecordActivity.java
@@ -14,23 +14,17 @@ package com.example.x15011071.audioacousticassistant_app;
  */
 
 import android.content.Intent;
-import android.os.Build;
-import android.os.Environment;
-import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
+import android.media.MediaPlayer;
+import android.media.MediaRecorder;
 import android.os.Bundle;
+import android.os.Environment;
+import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
-import android.media.MediaPlayer;
-import android.media.MediaPlayer.OnCompletionListener;
-import android.media.MediaRecorder;
-import android.util.Log;
-
 
 import java.io.File;
-import java.io.IOException;
 
 
 public class RecordActivity extends AppCompatActivity {
@@ -219,6 +213,7 @@ public class RecordActivity extends AppCompatActivity {
         intent.putExtras(d); //@reference Stack Overflow SetIntent & @authors
 
         startActivity(intent);
+        overridePendingTransition(R.anim.fade_in_two, R.anim.fade_out_two);
     }
 
     @Override //disables back button.
