--- 3. AudioAcousticAssistant-App/app/src/main/java/com/example/x15011071/audioacousticassistant_app/SplashActivity.java
@@ -1,63 +1,0 @@
-package com.example.x15011071.audioacousticassistant_app;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
-import android.view.KeyEvent;
-import android.widget.Toast;
-
-/**
- * Created by user on 16/03/2017.
- */
-
-public class SplashActivity extends AppCompatActivity {
-    private static int WELCOME_TIMEOUT = 3000;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        setContentView(R.layout.splash_screen);
-
-        new Handler().postDelayed(new Runnable(){
-            @Override
-            public void run(){
-                Intent welcome = new Intent(SplashActivity.this, MainActivity.class);
-                startActivity(welcome);
-                overridePendingTransition(R.anim.fade_in,R.anim.fade_out);
-                finish();
-            }
-        }, WELCOME_TIMEOUT);
-
-        Thread myThread = new Thread(){
-            @Override
-            public void run(){
-                try {
-                    sleep(3000);
-                    Intent intent = new Intent(getApplicationContext(), MainActivity.class);
-                    startActivity(intent);
-                    finish();
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                }
-            }
-
-        };
-        myThread.start();
-
-    }
-
-    @Override
-    public void onBackPressed() {
-        Toast.makeText(getApplicationContext(),"Sorry, you can't go back here",Toast.LENGTH_LONG).show();
-    }
-
-    public boolean onKeyDown(int keyCode, KeyEvent event) {
-        if (keyCode == KeyEvent.KEYCODE_VOLUME_DOWN
-                || keyCode == KeyEvent.KEYCODE_VOLUME_DOWN)
-            return true;
-        else
-            return false;
-    }
-}
